backend/pyproject.toml
[project]
name = "formbuilder-backend"
version = "0.1.0"
description = "FormBuilder API — LangGraph agent for electrical installation forms"
requires-python = ">=3.12"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "uvloop>=0.20.0",
    "httptools>=0.6.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "cloud-sql-python-connector[asyncpg]>=1.10.0",

    # LangGraph Agent
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",

    # Google Cloud
    "google-cloud-storage>=2.16.0",
    "google-cloud-secret-manager>=2.20.0",
    "gspread>=6.1.0",
    "google-auth>=2.29.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",

    # Utilities
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "orjson>=3.10.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
